+++ panel/views/field/row-fields/field.php
@@ -17,6 +17,7 @@ unset($sub['when']);
 	<?php $this->insert('field/field', [
 		'field' => $sub,
 		'ownLocales' => !$this->unwrap($ownsLocales ?? false),
+		'bareLabel' => !$this->unwrap($labels ?? true),
 		'data' => $fieldsData[$subName] ?? null,
 		'nameRoot' => "{$rowName}[fields][{$subName}]",
 		'idRoot' => "{$rowId}-{$subName}",